extract Chips and Trees
